Feature/circular deps (#5519)
* Start circular dependency work

* Resolve circular dependencies

- Add a Webpack plugin to our dev config that checks for circular dependencies
- Resolve circular dependencies involving resetEventsPolling (which was used inside the store
and itself relied on the store)
- Resolve Lish/Glish dependency chain (Glish was importing methods from Lish, which in turn
imported Glish)
- Resolve reportException chain (moved handleUnauthedErrors to its own utility)

* Update request interval from combineEventsMiddleware

* Abstract resetPollingInterval

* Only increment polling interval if it's below the maximum

* Refactor logic

* Review feedback 1

- Safe access Redux store in events.ts
- Remove circular-dep checker package from SDK
- Fix poll iteration conditional logic

* Fix broken imports

* Extract Sentry initialization

To resolve a circular dependency (reportException uses the store
to provide username etc. to Sentry, so importing it from within
the store was causing trouble), created src/initSentry, which
abstracts the initialization logic so that components within
the store can report errors to Sentry without causing a circular
import.

* Remove unused updateLinode action and updateByID utility method
