+++ packages/manager/package.json
@@ -203,6 +203,7 @@
     "browserslist": "^4.6.6",
     "case-sensitive-paths-webpack-plugin": "^2.2.0",
     "chalk": "^2.4.2",
+    "circular-dependency-plugin": "^5.2.0",
     "core-js": "^2.6.5",
     "coveralls": "^3.0.2",
     "css-loader": "^2.1.0",
