+++ packages/manager/src/features/linodes/LinodesLanding/LinodeActionMenu.tsx
@@ -12,7 +12,7 @@ import { compose } from 'recompose';
 
 import { RouteComponentProps, withRouter } from 'react-router-dom';
 
-import { lishLaunch } from 'src/features/Lish';
+import { lishLaunch } from 'src/features/Lish/lishUtils';
 
 import ActionMenu, { Action } from 'src/components/ActionMenu/ActionMenu';
