+++ packages/manager/src/store/events/event.actions.ts
@@ -1,5 +1,5 @@
 import actionCreatorFactory from 'typescript-fsa';
-import { ExtendedEvent } from './event.helpers';
+import { ExtendedEvent } from './event.types';
 
 type Event = ExtendedEvent;
 
@@ -12,3 +12,7 @@ export const actionCreator = actionCreatorFactory(`@@manager/events`);
 export const addEvents = actionCreator<Event[]>(ADD_EVENTS);
 
 export const updateEventsAsSeen = actionCreator(UPDATE_EVENTS_AS_SEEN);
+
+export const setPollingInterval = actionCreator<number>('set-polling-interval');
+
+export const setRequestDeadline = actionCreator<number>('set-request-deadline');
