+++ packages/manager/src/utilities/errorUtils.ts
@@ -2,8 +2,6 @@ import { APIError } from 'linode-js-sdk/lib/types';
 import { pathOr } from 'ramda';
 import { DEFAULT_ERROR_MESSAGE } from 'src/constants';
 
-import { reportException } from 'src/exceptionReporting';
-
 /**
  *
  * Override the default error message provided by our Axios
@@ -47,53 +45,6 @@ const isDefaultError = (errorResponse: APIError[]) => {
     errorResponse.length === 1 &&
     errorResponse[0].reason === DEFAULT_ERROR_MESSAGE
   );
-};
-
-export const handleUnauthorizedErrors = (
-  e: APIError[],
-  unauthedMessage: string
-) => {
-  /**
-   * filter out errors that match the following
-   * {
-   *   reason: "Unauthorized"
-   * }
-   *
-   * and if any of these errors exist, set the hasUnauthorizedError
-   * flag to true
-   */
-  let hasUnauthorizedError = false;
-  let filteredErrors: APIError[] = [];
-
-  try {
-    filteredErrors = e.filter(eachError => {
-      if (
-        typeof eachError.reason === 'string' &&
-        eachError.reason.toLowerCase().includes('unauthorized')
-      ) {
-        hasUnauthorizedError = true;
-        return false;
-      }
-      return true;
-    });
-  } catch (caughtError) {
-    reportException(`Error with Unauthed error handling: ${caughtError}`, {
-      apiError: e
-    });
-  }
-
-  /**
-   * if we found an unauthorized error, add on the new message in the API
-   * Error format
-   */
-  return hasUnauthorizedError
-    ? [
-        {
-          reason: unauthedMessage
-        },
-        ...filteredErrors
-      ]
-    : filteredErrors;
 };
 
 export const getErrorStringOrDefault = (
